+++ src/components/Landing/Landing.css
@@ -1,0 +1,25 @@
+.root{
+    padding: 10% 0;
+
+}
+
+.get-started-button {
+    background: linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%);
+    border-radius: 3px;
+    border: 0;
+    color: white;
+    height: 48px;
+    padding: 0 30px;
+    box-shadow: 0 3px 5px 2px rgba(255, 105, 135, .30);
+}
+
+img.bg {
+    min-height: 100%;
+    min-width: 1024px;
+    width: 100%;
+    height: auto;
+    z-index: -100;
+    position: fixed;
+    top: 0;
+    left: 0;
+}
