+++ src/components/Logo/Logo.css
@@ -1,0 +1,8 @@
+.Logo {
+    padding: 8px;
+    height: 50px;
+}
+
+.Logo img {
+    height: 100%;
+}
