+++ src/containers/Header/Header.js
@@ -1,0 +1,108 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import {withStyles} from 'material-ui/styles';
+import AppBar from 'material-ui/AppBar';
+import Toolbar from 'material-ui/Toolbar';
+import Typography from 'material-ui/Typography';
+import IconButton from 'material-ui/IconButton';
+import AccountCircle from 'material-ui-icons/AccountCircle';
+import Switch from 'material-ui/Switch';
+import {FormControlLabel, FormGroup} from 'material-ui/Form';
+import Menu, {MenuItem} from 'material-ui/Menu';
+import {Link} from 'react-router-dom'
+
+import Logo from '../../components/Logo/Logo'
+
+const styles = {
+    root: {
+        width: '100%',
+    },
+    flex: {
+        flex: 1,
+    },
+    menuButton: {
+        marginLeft: -12,
+        marginRight: 20,
+    },
+};
+
+class Header extends React.Component {
+    state = {
+        auth: true,
+        anchorEl: null,
+    };
+
+    handleChange = (event, checked) => {
+        this.setState({auth: checked});
+    };
+
+    handleMenu = event => {
+        this.setState({anchorEl: event.currentTarget});
+    };
+
+    handleClose = () => {
+        this.setState({anchorEl: null});
+    };
+
+    render() {
+        const {classes} = this.props;
+        const {auth, anchorEl} = this.state;
+        const open = Boolean(anchorEl);
+
+        return (
+            <div className={classes.root}>
+                <FormGroup>
+                </FormGroup>
+                <AppBar position="static">
+                    <Toolbar>
+                        <Link to="/"><Logo/></Link>
+                        <Typography type="title" color="inherit" className={classes.flex}>
+                            This is the header, you may modify it......
+                        </Typography>
+                        {auth && (
+                            <div>
+                                <IconButton
+                                    aria-owns={open ? 'menu-appbar' : null}
+                                    aria-haspopup="true"
+                                    onClick={this.handleMenu}
+                                    color="inherit"
+                                >
+                                    <AccountCircle/>
+                                </IconButton>
+                                <Menu
+                                    id="menu-appbar"
+                                    anchorEl={anchorEl}
+                                    anchorOrigin={{
+                                        vertical: 'top',
+                                        horizontal: 'right',
+                                    }}
+                                    transformOrigin={{
+                                        vertical: 'top',
+                                        horizontal: 'right',
+                                    }}
+                                    open={open}
+                                    onClose={this.handleClose}
+                                >
+                                    <MenuItem onClick={this.handleClose}>Profile</MenuItem>
+                                    <MenuItem onClick={this.handleClose}>My account</MenuItem>
+                                </Menu>
+                            </div>
+                        )}
+                        <FormControlLabel
+                            control={
+                                <Switch checked={auth} onChange={this.handleChange} aria-label="LoginSwitch"/>
+                            }
+                            label={auth ? 'Logout' : 'Login'}
+                        />
+                    </Toolbar>
+                </AppBar>
+            </div>
+        );
+    }
+}
+
+Header.propTypes = {
+    classes: PropTypes.object.isRequired,
+};
+
+export default withStyles(styles)(Header);
